requirements.txt
Flask
numpy
pandas
slackclient 
slackeventsapi
gunicorn
python-dotenv
requests